# openapi/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: API
paths:
...
  /users:
    get:
      summary: Get All Users
      operationId: GetUsers
      tags:
        - users
      responses:
        '200':
          description: All Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      
    post:
      summary: Create User
      operationId: CreateUser
      tags:
        - users
      requestBody:
        description: The User To Create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User Was Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}/tasks:
    get:
      summary: Get Tasks By User ID
      operationId: GetTasksByUserID
      tags:
        - users
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully Got Tasks By User ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
  /users/{id}:
    get:
      summary: Get User By ID
      operationId: GetUserByID
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User Was Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      
    patch:
      summary: Update User's password
      operationId: UpdateUserPassword
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: The New Password To Update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordRequest'
      responses:
        '200':
          description: The Password Was Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      
    put:
      summary: Update User
      operationId: UpdateUser
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: User Params For Update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User Params Was Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      
    delete:
      summary: Delete User By ID
      operationId: DeleteUserByID
      tags:
        - users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User Was Successfully Deleted
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        is_done:
          type: boolean
        user_id:
          type: string
          format: uuid
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        password:
          type: string
    UserRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    UpdateUserPasswordRequest:
      type: object
      properties:
        password:
          type: string